connecting business to handler an db

+++ business/entity/secret.go
@@ -3,5 +3,5 @@ package entity
 // Secret struct holds secret info
 type Secret struct {
 	SecretWord string
-	Number     int64
+	Number     int
 }
